refactor(agent): split oversized modules into focused files
- summarizer.py (784→323 lines): extract data tables to summarizer_data.py
- processing.py (427→282 lines): extract _symbol_to_graph to symbol_graph.py
- import_analyzer.py (674→26 lines): split into import_analyzers/ subpackage
  with per-language modules (python, go, typescript, rust, ruby)
- manifest_parser.py (470→23 lines): split into manifest_parsers/ subpackage
  with per-format parsers (npm, go, pip, pyproject, cargo)

Original modules remain as backward-compatible re-export facades.
No file exceeds 200 lines (largest: go_imports.py at 199).
