+++ agent/src/opentrace_agent/sources/code/import_analyzers/__init__.py
@@ -1,0 +1,49 @@
+# Copyright 2026 OpenTrace Contributors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Per-language import analysis subpackage — re-exports all public symbols."""
+
+from opentrace_agent.sources.code.import_analyzers.go_imports import (  # noqa: F401
+    analyze_go_imports,
+    reset_dir_index_cache,
+)
+from opentrace_agent.sources.code.import_analyzers.python_imports import (  # noqa: F401
+    analyze_python_imports,
+)
+from opentrace_agent.sources.code.import_analyzers.ruby_imports import (  # noqa: F401
+    analyze_ruby_imports,
+)
+from opentrace_agent.sources.code.import_analyzers.rust_imports import (  # noqa: F401
+    analyze_rust_imports,
+)
+from opentrace_agent.sources.code.import_analyzers.typescript_imports import (  # noqa: F401
+    analyze_typescript_imports,
+)
+from opentrace_agent.sources.code.import_analyzers.types import (  # noqa: F401
+    ImportResult,
+    package_id,
+    package_source_url,
+)
+
+__all__ = [
+    "ImportResult",
+    "analyze_go_imports",
+    "analyze_python_imports",
+    "analyze_ruby_imports",
+    "analyze_rust_imports",
+    "analyze_typescript_imports",
+    "package_id",
+    "package_source_url",
+    "reset_dir_index_cache",
+]
