+++ agent/src/opentrace_agent/sources/code/manifest_parsers/types.py
@@ -1,0 +1,38 @@
+# Copyright 2026 OpenTrace Contributors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Shared data types for manifest parsers."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+
+
+@dataclass
+class ParsedDependency:
+    """A single dependency declaration from a manifest file."""
+
+    name: str
+    version: str
+    registry: str  # "npm", "pypi", "go", "crates"
+    source: str  # manifest filename
+    dependency_type: str  # "runtime", "dev", "peer", "optional", "indirect"
+
+
+@dataclass
+class ManifestParseResult:
+    """Result of parsing a manifest file."""
+
+    dependencies: list[ParsedDependency] = field(default_factory=list)
+    errors: list[str] = field(default_factory=list)
