+++ agent/src/opentrace_agent/sources/code/summarizer.py
@@ -23,6 +23,14 @@ from __future__ import annotations
 import re
 from dataclasses import dataclass, field
 
+from opentrace_agent.sources.code.summarizer_data import (
+    CLASS_SUFFIX_MAP,
+    DIR_PATTERNS,
+    FILE_PATTERNS,
+    VERB_MAP,
+    _KEYWORD_PATTERNS,
+)
+
 
 # ---------------------------------------------------------------------------
 # Types
@@ -44,475 +52,6 @@ class SymbolMetadata:
     source: str | None = None
     docs: str | None = None
 
-
-# ---------------------------------------------------------------------------
-# Verb prefix → description mapping
-# ---------------------------------------------------------------------------
-
-VERB_MAP: dict[str, str] = {
-    "get": "Retrieves",
-    "fetch": "Retrieves",
-    "load": "Retrieves",
-    "find": "Retrieves",
-    "query": "Retrieves",
-    "read": "Reads",
-    "lookup": "Looks up",
-    "set": "Updates",
-    "update": "Updates",
-    "modify": "Updates",
-    "patch": "Updates",
-    "put": "Updates",
-    "validate": "Validates",
-    "verify": "Validates",
-    "check": "Validates",
-    "ensure": "Ensures",
-    "create": "Creates",
-    "make": "Creates",
-    "build": "Creates",
-    "generate": "Creates",
-    "new": "Creates",
-    "add": "Adds",
-    "insert": "Adds",
-    "append": "Adds",
-    "register": "Registers",
-    "delete": "Removes",
-    "remove": "Removes",
-    "destroy": "Removes",
-    "drop": "Removes",
-    "unset": "Removes",
-    "clear": "Clears",
-    "parse": "Parses",
-    "extract": "Extracts",
-    "decode": "Decodes",
-    "handle": "Handles",
-    "process": "Processes",
-    "run": "Runs",
-    "execute": "Executes",
-    "do": "Performs",
-    "perform": "Performs",
-    "convert": "Converts",
-    "transform": "Converts",
-    "to": "Converts to",
-    "format": "Formats",
-    "encode": "Encodes",
-    "serialize": "Serializes",
-    "marshal": "Serializes",
-    "deserialize": "Deserializes",
-    "unmarshal": "Deserializes",
-    "is": "Checks whether",
-    "has": "Checks whether",
-    "can": "Checks whether",
-    "should": "Checks whether",
-    "init": "Initializes",
-    "setup": "Initializes",
-    "configure": "Configures",
-    "start": "Starts",
-    "stop": "Stops",
-    "open": "Opens",
-    "close": "Closes",
-    "connect": "Connects",
-    "disconnect": "Disconnects",
-    "render": "Renders",
-    "display": "Renders",
-    "show": "Renders",
-    "draw": "Renders",
-    "paint": "Renders",
-    "hide": "Hides",
-    "send": "Sends",
-    "emit": "Emits",
-    "dispatch": "Dispatches",
-    "publish": "Publishes",
-    "broadcast": "Broadcasts",
-    "notify": "Notifies",
-    "receive": "Receives",
-    "listen": "Listens for",
-    "subscribe": "Subscribes to",
-    "on": "Handles",
-    "sort": "Sorts",
-    "filter": "Filters",
-    "map": "Maps",
-    "reduce": "Reduces",
-    "merge": "Merges",
-    "join": "Joins",
-    "split": "Splits",
-    "group": "Groups",
-    "flatten": "Flattens",
-    "test": "Tests",
-    "assert": "Tests",
-    "expect": "Tests",
-    "log": "Logs",
-    "print": "Logs",
-    "warn": "Logs warning for",
-    "error": "Logs error for",
-    "debug": "Logs debug info for",
-    "write": "Writes",
-    "save": "Saves",
-    "store": "Stores",
-    "cache": "Caches",
-    "flush": "Flushes",
-    "sync": "Synchronizes",
-    "reset": "Resets",
-    "refresh": "Refreshes",
-    "reload": "Reloads",
-    "retry": "Retries",
-    "wrap": "Wraps",
-    "unwrap": "Unwraps",
-    "apply": "Applies",
-    "resolve": "Resolves",
-    "reject": "Rejects",
-    "throw": "Throws",
-    "raise": "Raises",
-    "try": "Attempts",
-    "await": "Awaits",
-    "wait": "Waits for",
-    "schedule": "Schedules",
-    "defer": "Defers",
-    "cancel": "Cancels",
-    "abort": "Aborts",
-    "clone": "Clones",
-    "copy": "Copies",
-    "compare": "Compares",
-    "equals": "Checks equality of",
-    "match": "Matches",
-    "contains": "Checks whether contains",
-    "include": "Includes",
-    "exclude": "Excludes",
-    "enable": "Enables",
-    "disable": "Disables",
-    "toggle": "Toggles",
-    "mount": "Mounts",
-    "unmount": "Unmounts",
-    "use": "Uses",
-    "with": "Configures with",
-    "from": "Creates from",
-    "of": "Creates",
-}
-
-# ---------------------------------------------------------------------------
-# Class suffix patterns
-# ---------------------------------------------------------------------------
-
-CLASS_SUFFIX_MAP: dict[str, str] = {
-    "service": "Service",
-    "handler": "Handler",
-    "controller": "Controller",
-    "factory": "Factory",
-    "repository": "Repository",
-    "repo": "Repository",
-    "manager": "Manager",
-    "provider": "Provider",
-    "adapter": "Adapter",
-    "middleware": "Middleware",
-    "guard": "Guard",
-    "interceptor": "Interceptor",
-    "resolver": "Resolver",
-    "validator": "Validator",
-    "builder": "Builder",
-    "parser": "Parser",
-    "formatter": "Formatter",
-    "converter": "Converter",
-    "serializer": "Serializer",
-    "client": "Client",
-    "server": "Server",
-    "router": "Router",
-    "store": "Store",
-    "cache": "Cache",
-    "queue": "Queue",
-    "pool": "Pool",
-    "registry": "Registry",
-    "observer": "Observer",
-    "emitter": "Emitter",
-    "listener": "Listener",
-    "subscriber": "Subscriber",
-    "publisher": "Publisher",
-    "component": "Component",
-    "module": "Module",
-    "plugin": "Plugin",
-    "helper": "Helper",
-    "util": "Utility",
-    "utils": "Utility",
-    "error": "Error",
-    "exception": "Exception",
-    "model": "Model",
-    "entity": "Entity",
-    "dto": "DTO",
-    "config": "Configuration",
-    "options": "Options",
-    "context": "Context",
-    "state": "State",
-    "hook": "Hook",
-}
-
-# ---------------------------------------------------------------------------
-# File / directory pattern maps
-# ---------------------------------------------------------------------------
-
-FILE_PATTERNS: list[tuple[re.Pattern[str], str]] = [
-    (re.compile(r"_test\.go$"), "Tests for"),
-    (re.compile(r"_test\.py$"), "Tests for"),
-    (re.compile(r"\.test\.[jt]sx?$"), "Tests for"),
-    (re.compile(r"\.spec\.[jt]sx?$"), "Tests for"),
-    (re.compile(r"^test_"), "Tests for"),
-    (re.compile(r"^conftest\.py$"), "Pytest fixtures and configuration"),
-    (re.compile(r"^setup\.[jt]s$"), "Setup configuration"),
-    (re.compile(r"^index\.[jt]sx?$"), "Barrel exports for"),
-    (re.compile(r"^main\.[a-z]+$"), "Application entry point"),
-    (re.compile(r"^mod\.rs$"), "Module declarations"),
-    (re.compile(r"^__init__\.py$"), "Package initialization for"),
-    (re.compile(r"^constants?\.[a-z]+$"), "Constants and configuration values"),
-    (re.compile(r"^types?\.[a-z]+$"), "Type definitions"),
-    (re.compile(r"^utils?\.[a-z]+$"), "Utility functions"),
-    (re.compile(r"^helpers?\.[a-z]+$"), "Helper functions"),
-    (re.compile(r"^middleware\.[a-z]+$"), "Middleware definitions"),
-    (re.compile(r"^routes?\.[a-z]+$"), "Route definitions"),
-    (re.compile(r"^models?\.[a-z]+$"), "Data model definitions"),
-    (re.compile(r"^schema\.[a-z]+$"), "Schema definitions"),
-    (re.compile(r"^migrations?"), "Database migration"),
-    (re.compile(r"^dockerfile", re.IGNORECASE), "Docker container configuration"),
-    (re.compile(r"^makefile$", re.IGNORECASE), "Build automation rules"),
-    (re.compile(r"^readme", re.IGNORECASE), "Project documentation"),
-    (re.compile(r"^changelog", re.IGNORECASE), "Version change history"),
-    (re.compile(r"^license", re.IGNORECASE), "License information"),
-    (re.compile(r"\.config\.[a-z]+$"), "Configuration for"),
-    (re.compile(r"rc\.[a-z]+$"), "Configuration for"),
-]
-
-DIR_PATTERNS: dict[str, str] = {
-    "api": "API layer",
-    "apis": "API layer",
-    "handlers": "Request handlers",
-    "handler": "Request handlers",
-    "controllers": "Request controllers",
-    "controller": "Request controllers",
-    "routes": "Route definitions",
-    "routing": "Route definitions",
-    "models": "Data models",
-    "model": "Data models",
-    "entities": "Data entities",
-    "schema": "Schema definitions",
-    "schemas": "Schema definitions",
-    "services": "Service layer",
-    "service": "Service layer",
-    "middleware": "Middleware",
-    "middlewares": "Middleware",
-    "utils": "Utility functions",
-    "util": "Utility functions",
-    "helpers": "Helper functions",
-    "helper": "Helper functions",
-    "lib": "Library modules",
-    "libs": "Library modules",
-    "pkg": "Package modules",
-    "internal": "Internal packages",
-    "cmd": "Command entry points",
-    "config": "Configuration",
-    "configs": "Configuration",
-    "tests": "Test suite",
-    "test": "Test suite",
-    "__tests__": "Test suite",
-    "spec": "Test specifications",
-    "specs": "Test specifications",
-    "fixtures": "Test fixtures",
-    "mocks": "Test mocks",
-    "components": "UI components",
-    "component": "UI components",
-    "pages": "Page components",
-    "views": "View components",
-    "layouts": "Layout components",
-    "hooks": "React hooks",
-    "store": "State management",
-    "stores": "State management",
-    "state": "State management",
-    "reducers": "State reducers",
-    "actions": "State actions",
-    "selectors": "State selectors",
-    "types": "Type definitions",
-    "interfaces": "Interface definitions",
-    "constants": "Constants",
-    "static": "Static assets",
-    "assets": "Static assets",
-    "public": "Public assets",
-    "styles": "Stylesheets",
-    "css": "Stylesheets",
-    "docs": "Documentation",
-    "doc": "Documentation",
-    "scripts": "Build and utility scripts",
-    "migrations": "Database migrations",
-    "seeds": "Database seed data",
-    "templates": "Templates",
-    "i18n": "Internationalization",
-    "locales": "Locale translations",
-    "proto": "Protocol buffer definitions",
-    "generated": "Generated code",
-    "gen": "Generated code",
-    "dist": "Build output",
-    "build": "Build output",
-    "vendor": "Vendored dependencies",
-    "node_modules": "NPM dependencies",
-    "bin": "Executable binaries",
-    "examples": "Usage examples",
-    "example": "Usage examples",
-    "plugins": "Plugin modules",
-    "extensions": "Extension modules",
-    "auth": "Authentication and authorization",
-    "security": "Security modules",
-    "crypto": "Cryptographic utilities",
-    "db": "Database layer",
-    "database": "Database layer",
-    "cache": "Caching layer",
-    "queue": "Message queue handlers",
-    "workers": "Background workers",
-    "jobs": "Background jobs",
-    "tasks": "Task definitions",
-    "events": "Event handlers",
-    "subscribers": "Event subscribers",
-    "publishers": "Event publishers",
-    "adapters": "Adapter implementations",
-    "providers": "Service providers",
-    "repositories": "Data repositories",
-    "repository": "Data repositories",
-    "clients": "External API clients",
-    "sdk": "SDK modules",
-    "common": "Shared common modules",
-    "shared": "Shared modules",
-    "core": "Core modules",
-    "base": "Base classes and interfaces",
-    "errors": "Error definitions",
-    "exceptions": "Exception definitions",
-    "validators": "Input validators",
-    "validation": "Validation logic",
-    "serializers": "Data serializers",
-    "parsers": "Data parsers",
-    "formatters": "Data formatters",
-    "converters": "Data converters",
-    "transformers": "Data transformers",
-    "mappers": "Data mappers",
-    "resolvers": "GraphQL resolvers",
-    "guards": "Route guards",
-    "interceptors": "Request interceptors",
-    "decorators": "Decorators",
-    "annotations": "Annotations",
-    "factories": "Factory functions",
-    "builders": "Builder patterns",
-    "observers": "Observer implementations",
-    "strategies": "Strategy pattern implementations",
-    "commands": "Command implementations",
-    "queries": "Query implementations",
-    "notifications": "Notification handlers",
-    "mailers": "Email sending",
-    "logging": "Logging configuration",
-    "monitoring": "Monitoring and metrics",
-    "metrics": "Application metrics",
-    "tracing": "Distributed tracing",
-    "indexer": "Indexing pipeline",
-    "graph": "Graph data structures",
-    "summarizer": "Summarization modules",
-    "embedder": "Embedding modules",
-    "extractors": "Data extractors",
-}
-
-# ---------------------------------------------------------------------------
-# Keyword extraction patterns
-# ---------------------------------------------------------------------------
-
-_KEYWORD_PATTERNS: list[dict[str, object]] = [
-    {"domain": "database", "patterns": [
-        re.compile(r"\bsql\b", re.IGNORECASE), re.compile(r"\bquery\b", re.IGNORECASE),
-        re.compile(r"\bSELECT\b"), re.compile(r"\bINSERT\b"),
-        re.compile(r"\bUPDATE\b.*\bSET\b"), re.compile(r"\bdb\.", re.IGNORECASE),
-        re.compile(r"\bcursor\b", re.IGNORECASE), re.compile(r"\bconnection\b", re.IGNORECASE),
-        re.compile(r"\bprisma\b", re.IGNORECASE), re.compile(r"\bknex\b", re.IGNORECASE),
-        re.compile(r"\bsequelize\b", re.IGNORECASE), re.compile(r"\bmongoose\b", re.IGNORECASE),
-        re.compile(r"\bsqlx\b", re.IGNORECASE), re.compile(r"\borm\b", re.IGNORECASE),
-        re.compile(r"\bmigration\b", re.IGNORECASE), re.compile(r"\bschema\b", re.IGNORECASE),
-        re.compile(r"\btable\b", re.IGNORECASE), re.compile(r"\btransaction\b", re.IGNORECASE),
-        re.compile(r"\bcolumn\b", re.IGNORECASE),
-    ]},
-    {"domain": "auth", "patterns": [
-        re.compile(r"\bauth\b", re.IGNORECASE), re.compile(r"\bjwt\b", re.IGNORECASE),
-        re.compile(r"\btoken\b", re.IGNORECASE), re.compile(r"\bpassword\b", re.IGNORECASE),
-        re.compile(r"\blogin\b", re.IGNORECASE), re.compile(r"\bsession\b", re.IGNORECASE),
-        re.compile(r"\boauth\b", re.IGNORECASE), re.compile(r"\bbcrypt\b", re.IGNORECASE),
-        re.compile(r"\bcredential\b", re.IGNORECASE), re.compile(r"\bpermission\b", re.IGNORECASE),
-        re.compile(r"\brole\b", re.IGNORECASE), re.compile(r"\bacl\b", re.IGNORECASE),
-        re.compile(r"\bsignin\b", re.IGNORECASE), re.compile(r"\bsignup\b", re.IGNORECASE),
-    ]},
-    {"domain": "http", "patterns": [
-        re.compile(r"\bhttp\b", re.IGNORECASE), re.compile(r"\brequest\b", re.IGNORECASE),
-        re.compile(r"\bresponse\b", re.IGNORECASE), re.compile(r"\bfetch\s*\(", re.IGNORECASE),
-        re.compile(r"\baxios\b", re.IGNORECASE), re.compile(r"\bhandler\b", re.IGNORECASE),
-        re.compile(r"\bmiddleware\b", re.IGNORECASE), re.compile(r"\bcors\b", re.IGNORECASE),
-        re.compile(r"\bendpoint\b", re.IGNORECASE), re.compile(r"\broute\b", re.IGNORECASE),
-        re.compile(r"\bREST\b", re.IGNORECASE), re.compile(r"\bgraphql\b", re.IGNORECASE),
-        re.compile(r"\bheader\b", re.IGNORECASE),
-    ]},
-    {"domain": "filesystem", "patterns": [
-        re.compile(r"\bfs\.", re.IGNORECASE), re.compile(r"\breadFile\b", re.IGNORECASE),
-        re.compile(r"\bwriteFile\b", re.IGNORECASE), re.compile(r"\bpath\.", re.IGNORECASE),
-        re.compile(r"\bstream\b", re.IGNORECASE), re.compile(r"\bbuffer\b", re.IGNORECASE),
-        re.compile(r"\bmkdir\b", re.IGNORECASE), re.compile(r"\bunlink\b", re.IGNORECASE),
-        re.compile(r"\bglob\b", re.IGNORECASE),
-    ]},
-    {"domain": "crypto", "patterns": [
-        re.compile(r"\bcrypto\b", re.IGNORECASE), re.compile(r"\bencrypt\b", re.IGNORECASE),
-        re.compile(r"\bdecrypt\b", re.IGNORECASE), re.compile(r"\bhash\b", re.IGNORECASE),
-        re.compile(r"\bsign\b", re.IGNORECASE), re.compile(r"\bverify\b", re.IGNORECASE),
-        re.compile(r"\bcipher\b", re.IGNORECASE), re.compile(r"\bhmac\b", re.IGNORECASE),
-    ]},
-    {"domain": "cache", "patterns": [
-        re.compile(r"\bcache\b", re.IGNORECASE), re.compile(r"\bredis\b", re.IGNORECASE),
-        re.compile(r"\bmemcached\b", re.IGNORECASE), re.compile(r"\bttl\b", re.IGNORECASE),
-        re.compile(r"\binvalidate\b", re.IGNORECASE), re.compile(r"\bLRU\b"),
-    ]},
-    {"domain": "queue", "patterns": [
-        re.compile(r"\bqueue\b", re.IGNORECASE), re.compile(r"\bworker\b", re.IGNORECASE),
-        re.compile(r"\bjob\b", re.IGNORECASE), re.compile(r"\bpublish\b", re.IGNORECASE),
-        re.compile(r"\bsubscribe\b", re.IGNORECASE), re.compile(r"\bkafka\b", re.IGNORECASE),
-        re.compile(r"\brabbitmq\b", re.IGNORECASE), re.compile(r"\bamqp\b", re.IGNORECASE),
-    ]},
-    {"domain": "config", "patterns": [
-        re.compile(r"\bconfig\b", re.IGNORECASE), re.compile(r"\benv\b", re.IGNORECASE),
-        re.compile(r"\bsettings\b", re.IGNORECASE), re.compile(r"\byaml\b", re.IGNORECASE),
-        re.compile(r"\bdotenv\b", re.IGNORECASE), re.compile(r"\.env\b", re.IGNORECASE),
-    ]},
-    {"domain": "logging", "patterns": [
-        re.compile(r"\blogger\b", re.IGNORECASE), re.compile(r"\bwinston\b", re.IGNORECASE),
-        re.compile(r"\bpino\b", re.IGNORECASE), re.compile(r"\blogrus\b", re.IGNORECASE),
-        re.compile(r"\bslog\b", re.IGNORECASE), re.compile(r"\blog\.", re.IGNORECASE),
-    ]},
-    {"domain": "testing", "patterns": [
-        re.compile(r"\btest\b", re.IGNORECASE), re.compile(r"\bassert\b", re.IGNORECASE),
-        re.compile(r"\bexpect\b", re.IGNORECASE), re.compile(r"\bmock\b", re.IGNORECASE),
-        re.compile(r"\bstub\b", re.IGNORECASE), re.compile(r"\bspy\b", re.IGNORECASE),
-        re.compile(r"\bfixture\b", re.IGNORECASE), re.compile(r"\bjest\b", re.IGNORECASE),
-        re.compile(r"\bpytest\b", re.IGNORECASE),
-    ]},
-    {"domain": "async", "patterns": [
-        re.compile(r"\basync\b", re.IGNORECASE), re.compile(r"\bawait\b", re.IGNORECASE),
-        re.compile(r"\bPromise\b"), re.compile(r"\bgoroutine\b", re.IGNORECASE),
-        re.compile(r"\bchannel\b", re.IGNORECASE), re.compile(r"\bconcurrent\b", re.IGNORECASE),
-        re.compile(r"\bparallel\b", re.IGNORECASE), re.compile(r"\bmutex\b", re.IGNORECASE),
-    ]},
-    {"domain": "error", "patterns": [
-        re.compile(r"\btry\b"), re.compile(r"\bcatch\b"), re.compile(r"\bthrow\b"),
-        re.compile(r"\bError\b"), re.compile(r"\bpanic\b", re.IGNORECASE),
-        re.compile(r"\brecover\b", re.IGNORECASE), re.compile(r"\berrno\b", re.IGNORECASE),
-    ]},
-    {"domain": "validation", "patterns": [
-        re.compile(r"\bvalidate\b", re.IGNORECASE), re.compile(r"\bsanitize\b", re.IGNORECASE),
-        re.compile(r"\bzod\b", re.IGNORECASE), re.compile(r"\byup\b", re.IGNORECASE),
-        re.compile(r"\bjoi\b", re.IGNORECASE), re.compile(r"\bregexp\b", re.IGNORECASE),
-        re.compile(r"\bregex\b", re.IGNORECASE),
-    ]},
-    {"domain": "websocket", "patterns": [
-        re.compile(r"\bwebsocket\b", re.IGNORECASE), re.compile(r"\bws\.", re.IGNORECASE),
-        re.compile(r"\bsocket\b", re.IGNORECASE), re.compile(r"\bWebSocket\b"),
-        re.compile(r"\bemit\s*\(", re.IGNORECASE),
-    ]},
-    {"domain": "email", "patterns": [
-        re.compile(r"\bemail\b", re.IGNORECASE), re.compile(r"\bsmtp\b", re.IGNORECASE),
-        re.compile(r"\bsendmail\b", re.IGNORECASE), re.compile(r"\bmailer\b", re.IGNORECASE),
-    ]},
-]
 
 
 # ---------------------------------------------------------------------------
